+++ RELEASES.md
@@ -13,6 +13,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-metrics: added traffic lights
 * service-metrics: computing composite traffic lights
 * service-metrics: rating report
+* prov-template-compiler: generated code for jsweet/jsweet.sh to ensure compatibility with latest jdk
 
 `2.2.0`
 * docker-service-templates: deployed dictionary for explanations, build for intel and arm platforms